Create TComments type
