+++ src/lib/hooks/useDeleteComment.ts
@@ -10,20 +10,14 @@ import {
 } from '@tanstack/react-query';
 
 import { QueryKey } from '../enums';
-import { type CommentWithRelationsAndRelationCountsAndUserReaction } from '../types';
+import type {
+  CommentWithRelationsAndRelationCountsAndUserReaction,
+  TComments,
+} from '../types';
 
 type TComment = {
   data: { comment: Comment | null } | null;
   errors: { [key: string]: string[] } | unknown | null;
-  success: boolean;
-};
-
-type TComments = {
-  data: {
-    comments: CommentWithRelationsAndRelationCountsAndUserReaction[];
-    nextCursor: number | null;
-  };
-  errors: { [key: string]: string[] } | null;
   success: boolean;
 };
