+++ src/lib/hooks/useDeleteReaction.ts
@@ -13,18 +13,10 @@ import { QueryKey } from '../enums';
 import type {
   CommentWithRelationsAndRelationCountsAndUserReaction,
   PostWithRelationsAndRelationCountsAndUserReaction,
+  TComments,
   TPost,
   TPosts,
 } from '../types';
-
-type TComments = {
-  data: {
-    comments: CommentWithRelationsAndRelationCountsAndUserReaction[];
-    nextCursor: number | null;
-  };
-  errors: { [key: string]: string[] } | null;
-  success: boolean;
-};
 
 type TReaction = {
   data: { reaction: Reaction | null } | null;
